Crate x448

RustCrypto: X448 Diffie-Hellman

THIS CODE HAS NOT BEEN AUDITED OR REVIEWED. USE AT YOUR OWN RISK.

About

This crate provides a pure Rust implementation of X448, the Diffie-Hellman function for the Ed448-Goldilocks elliptic curve, providing public key cryptography for symmetric key exchange/agreement.

Companion library to ed448-goldilocks.

License

All crates licensed under either of

Apache License, Version 2.0

MIT license

at your option.

Contribution

Unless you explicitly state otherwise, any contribution intentionally submitted for inclusion in the work by you, as defined in the Apache-2.0 license, shall be dual licensed as above, without any additional terms or conditions.

Structs

EphemeralSecret

An EphemeralSecret is a Scalar on Curve448.

PublicKey

A PublicKey is a point on Curve448.

SharedSecret

A SharedSecret is a point on Curve448. This point is the result of a Diffie-Hellman key exchange.

StaticSecret

A Diffie-Hellman secret key that can be used to compute multiple SharedSecrets.

Constants

X448_BASEPOINT_BYTES

Functions

slice_to_array 🔒

x448

A safe version of the x448 function defined in RFC448. Currently, the only reason I can think of for using the raw function is FFI. Option is FFI safe1. So we can still maintain that the invariant that we do not return a low order point.

x448_unchecked

An unchecked version of the x448 function defined in RFC448 No checks are made on the points.

Type Aliases

EphemeralSecretBytes 🔒

MontgomeryScalar 🔒