x448

Function x448

pub fn x448(scalar_bytes: [u8; 56], point_bytes: [u8; 56]) -> Option<[u8; 56]>

A safe version of the x448 function defined in RFC448. Currently, the only reason I can think of for using the raw function is FFI. Option is FFI safe1. So we can still maintain that the invariant that we do not return a low order point.